+++ provisioners/bash/rhel7.sh
@@ -23,8 +23,13 @@ yum -y upgrade --exclude=python-urllib3-*
 # yum -y install puppet-agent epel-release
 rpm -ivh --force "https://dl.fedoraproject.org/pub/epel/epel-release-latest-${OS_VERSION}.noarch.rpm"
 
-yum -y install "https://yum.puppetlabs.com/puppet${PUPPET_MAJOR_VERSION}/${OS_TYPE}/${OS_VERSION}/${ARCH_TYPE}/puppet-agent-${PUPPET_AGENT_VERSION}-1.${OS_TYPE}${OS_VERSION}.${ARCH_TYPE}.rpm"
-
+if [[ "$(/opt/puppetlabs/puppet/bin/puppet --version)" == "${PUPPET_AGENT_VERSION}" ]]
+then
+    echo "Puppet is already installed. Skipping Puppet installation..."
+else
+    echo "Installing puppet..."
+    yum -y install "https://yum.puppetlabs.com/puppet${PUPPET_MAJOR_VERSION}/${OS_TYPE}/${OS_VERSION}/${ARCH_TYPE}/puppet-agent-${PUPPET_AGENT_VERSION}-1.${OS_TYPE}${OS_VERSION}.${ARCH_TYPE}.rpm"
+fi
 yum-config-manager --enable rhui-REGION-rhel-server-optional
 yum-config-manager --enable rhui-REGION-rhel-server-extras
 yum-config-manager --enable "rhel-${OS_VERSION}-server-rhui-optional-rpms"