+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add aws.install_rng_tools configuration for optimising entropy on instance types which use NVMe
+- Avoid overwriting Puppet installation when one is already installed in the source/base AMI [#241]
 
 ### Changed
 - Lock down pylint to 2.6.0